+++ ksyun/resource_ksyun_redis_instance_test.go
@@ -23,6 +23,12 @@ func TestAccKcs_basic(t *testing.T) {
 					testAccCheckKcsInstanceExists("ksyun_redis_instance.default"),
 				),
 			},
+			{
+				Config: testUpdateAccKcsConfig,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckKcsInstanceExists("ksyun_redis_instance.default"),
+				),
+			},
 		},
 	})
 }
@@ -68,7 +74,7 @@ func testAccCheckKcsDestroy(s *terraform.State) error {
 
 const testAccKcsConfig = `
 variable "available_zone" {
-  default = "cn-beijing-6a"
+  default = "cn-shanghai-2b"
 }
 
 variable "protocol" {
@@ -80,7 +86,7 @@ resource "ksyun_vpc" "default" {
   cidr_block = "10.7.0.0/21"
 }
 
-resource "ksyun_subnet" "foo" {
+resource "ksyun_subnet" "default" {
   subnet_name      	= "ksyun-subnet-tf"
   cidr_block 		= "10.7.0.0/21"
   subnet_type 		= "Reserve"
@@ -114,6 +120,72 @@ resource "ksyun_redis_instance" "default" {
   pass_word             = "Shiwo1101"
   iam_project_id        = "0"
   slave_num             = 0  
+  protocol              = "${var.protocol}"
+  reset_all_parameters  = false
+  parameters = {
+    "appendonly"                  = "no",
+    "appendfsync"                 = "everysec",
+    "maxmemory-policy"            = "volatile-lru",
+    "hash-max-ziplist-entries"    = "512",
+    "zset-max-ziplist-entries"    = "128",
+    "list-max-ziplist-size"       = "-2",
+    "hash-max-ziplist-value"      = "64",
+    "notify-keyspace-events"      = "",
+    "zset-max-ziplist-value"      = "64",
+    "maxmemory-samples"           = "5",
+    "set-max-intset-entries"      = "512",
+    "timeout"                     = "600",
+  }
+}
+`
+const testUpdateAccKcsConfig = `
+variable "available_zone" {
+  default = "cn-shanghai-2b"
+}
+
+variable "protocol" {
+  default = "4.0"
+}
+
+resource "ksyun_vpc" "default" {
+  vpc_name   = "ksyun-vpc-tf"
+  cidr_block = "10.7.0.0/21"
+}
+
+resource "ksyun_subnet" "default" {
+  subnet_name      	= "ksyun-subnet-tf"
+  cidr_block 		= "10.7.0.0/21"
+  subnet_type 		= "Reserve"
+  dhcp_ip_from 		= "10.7.0.2"
+  dhcp_ip_to 		= "10.7.0.253"
+  vpc_id 			= "${ksyun_vpc.default.id}"
+  gateway_ip 		= "10.7.0.1"
+  dns1 				= "198.18.254.41"
+  dns2 				= "198.18.254.40"
+  availability_zone = "${var.available_zone}"
+}
+
+resource "ksyun_redis_sec_group" "default" {
+  available_zone 	= "${var.available_zone}"
+  name 				= "testTerraform000"
+  description 		= "testTerraform000"
+}
+
+resource "ksyun_redis_instance" "default" {
+  available_zone        = "${var.available_zone}"
+  name                  = "MyRedisInstance1101"
+  mode                  = 2
+  capacity              = 1
+  net_type              = 2
+  security_group_id     = "${ksyun_redis_sec_group.default.id}"
+  vnet_id 				= "${ksyun_subnet.default.id}"
+  vpc_id 				= "${ksyun_vpc.default.id}"
+  bill_type             = 5
+  duration              = ""
+  duration_unit         = ""
+  pass_word             = "wwsNewPwd123"
+  iam_project_id        = "0"
+  slave_num             = 0
   protocol              = "${var.protocol}"
   reset_all_parameters  = false
   parameters = {